Reorganize code and add comments
