+++ crates/noseyparker/src/datastore.rs
@@ -7,6 +7,7 @@ use std::path::{Path, PathBuf};
 use tracing::{debug, debug_span};
 
 use crate::blob_id::BlobId;
+use crate::git_url::GitUrl;
 use crate::location::{Location, OffsetSpan, SourcePoint, SourceSpan};
 use crate::match_type::Match;
 use crate::provenance::Provenance;
@@ -15,6 +16,14 @@ use crate::snippet::Snippet;
 // -------------------------------------------------------------------------------------------------
 // Datastore
 // -------------------------------------------------------------------------------------------------
+
+/// The source of truth for Nosey Parker findings and runtime state.
+///
+/// A `Datastore` resides on disk as a directory, and stores a number of things:
+///
+/// - A sqlite database for recording findings and scan information
+/// - A scratch directory for providing temporary directories and files
+/// - A directory used for storing clones of Git repositories
 pub struct Datastore {
     /// The root directory of everything contained in this `Datastore`.
     root_dir: PathBuf,
@@ -23,6 +32,7 @@ pub struct Datastore {
     conn: Connection,
 }
 
+// Public implementation
 impl Datastore {
     /// Create a new datastore at `root_dir` if one does not exist,
     /// or open an existing one if present.
@@ -92,6 +102,210 @@ impl Datastore {
         self.root_dir.join("clones")
     }
 
+    /// Get a path for a local clone of the given git URL within this datastore's clones directory.
+    pub fn clone_destination(&self, repo: &GitUrl) -> Result<std::path::PathBuf> {
+        clone_destination(&self.clones_dir(), repo)
+    }
+
+    /// Analyze the datastore's sqlite database, potentially allowing for better query planning
+    pub fn analyze(&self) -> Result<()> {
+        self.conn.execute("analyze", [])?;
+        Ok(())
+    }
+
+    /// Record the given matches into the datastore.
+    pub fn record_matches<'a, T: IntoIterator<Item = &'a Match>>(
+        &mut self,
+        matches: T,
+    ) -> Result<usize> {
+        let _span = debug_span!("Datastore::record_matches", "{}", self.root_dir.display()).entered();
+
+        let tx = self.conn.transaction()?;
+        let mut stmt = tx.prepare_cached(indoc! {r#"
+            insert or replace into matches(
+                blob_id,
+                start_byte,
+                end_byte,
+                start_line,
+                start_column,
+                end_line,
+                end_column,
+                before_snippet,
+                matching_input,
+                after_snippet,
+                group_index,
+                group_input,
+                rule_name,
+                provenance_type,
+                provenance
+            ) values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+        "#})?;
+        let mut num_changed = 0;
+        for m in matches {
+            let span = &m.location.offset_span;
+            let src = &m.location.source_span;
+            let (ptype, ppath) = match &m.provenance {
+                Provenance::File { path } => ("file", path.to_string_lossy()),
+                Provenance::GitRepo { path } => ("git", path.to_string_lossy()),
+            };
+            // FIXME: the number of changed rows is not the number of newly found matches!
+            num_changed += stmt.execute((
+                m.blob_id.hex(),
+                span.start,
+                span.end,
+                src.start.line,
+                src.start.column,
+                src.end.line,
+                src.end.column,
+                m.snippet.before.as_slice(),
+                m.snippet.matching.as_slice(),
+                m.snippet.after.as_slice(),
+                &m.capture_group_index,
+                m.match_content.as_slice(),
+                &m.rule_name,
+                ptype,
+                ppath,
+            ))?;
+        }
+        drop(stmt);
+        tx.commit()?;
+        Ok(num_changed)
+    }
+
+    /// Summarize all recorded findings.
+    pub fn summarize(&self) -> Result<MatchSummary> {
+        let _span = debug_span!("Datastore::summarize", "{}", self.root_dir.display()).entered();
+
+        let mut stmt = self.conn.prepare_cached(indoc! {r#"
+            select rule_name, count(*) grouped_count, sum(num_matches) total_count
+            from (
+                select group_input, rule_name, count(*) num_matches
+                from matches
+                group by 1, 2
+            )
+            group by 1
+            order by grouped_count desc
+        "#})?;
+        let entries = stmt.query_map((), |row| {
+            Ok(MatchSummaryEntry {
+                rule_name: row.get(0)?,
+                distinct_count: row.get(1)?,
+                total_count: row.get(2)?,
+            })
+        })?;
+        let mut es = Vec::new();
+        for e in entries {
+            es.push(e?);
+        }
+        Ok(MatchSummary(es))
+    }
+
+    /// Get the root directory that contains this `Datastore`.
+    pub fn root_dir(&self) -> &Path {
+        &self.root_dir
+    }
+
+    /// Get metadata for all groups of identical matches recorded within this `Datastore`.
+    pub fn get_match_group_metadata(&self) -> Result<Vec<MatchGroupMetadata>> {
+        let _span =
+            debug_span!("Datastore::get_match_group_metadata", "{}", self.root_dir.display()).entered();
+
+        let mut stmt = self.conn.prepare_cached(indoc! {r#"
+            select group_input, rule_name, count(*)
+            from matches
+            group by 1, 2
+            order by 2
+        "#})?;
+        let entries = stmt.query_map((), |row| {
+            Ok(MatchGroupMetadata {
+                match_content: BString::new(row.get(0)?),
+                rule_name: row.get(1)?,
+                num_matches: row.get(2)?,
+            })
+        })?;
+        let mut es = Vec::new();
+        for e in entries {
+            es.push(e?);
+        }
+        Ok(es)
+    }
+
+    /// Get up to `limit` matches that belong to the group with the given group metadata.
+    pub fn get_match_group_matches(
+        &self,
+        metadata: &MatchGroupMetadata,
+        limit: Option<usize>,
+    ) -> Result<Vec<Match>> {
+        let _span = debug_span!("Datastore::match_groups", "{}", self.root_dir.display()).entered();
+
+        let mut stmt = self.conn.prepare_cached(indoc! {r#"
+            select
+                blob_id,
+                start_byte,
+                end_byte,
+                start_line,
+                start_column,
+                end_line,
+                end_column,
+                before_snippet,
+                matching_input,
+                after_snippet,
+                group_index,
+                provenance_type,
+                provenance
+            from matches
+            where rule_name = ? and group_input = ?
+            order by blob_id, start_byte, end_byte
+            limit ?
+        "#})?;
+
+        let limit: i64 = match limit {
+            Some(limit) => limit.try_into().expect("limit should be convertible"),
+            None => -1,
+        };
+        let entries = stmt.query_map((&metadata.rule_name, metadata.match_content.as_slice(), limit), |row| {
+            let v0: String = row.get(0)?;
+            Ok(Match {
+                blob_id: BlobId::from_hex(&v0).expect("blob id from database should be valid"),
+                location: Location {
+                    offset_span: OffsetSpan {
+                        start: row.get(1)?,
+                        end: row.get(2)?,
+                    },
+                    source_span: SourceSpan {
+                        start: SourcePoint {
+                            line: row.get(3)?,
+                            column: row.get(4)?,
+                        },
+                        end: SourcePoint {
+                            line: row.get(5)?,
+                            column: row.get(6)?,
+                        },
+                    },
+                },
+                snippet: Snippet {
+                    before: BString::new(row.get(7)?),
+                    matching: BString::new(row.get(8)?),
+                    after: BString::new(row.get(9)?),
+                },
+                capture_group_index: row.get(10)?,
+                match_content: metadata.match_content.clone(),
+                rule_name: metadata.rule_name.clone(),
+                provenance: provenance_from_parts(row.get(11)?, row.get(12)?)
+                    .expect("provenance value from database should be valid"),
+            })
+        })?;
+        let mut es = Vec::new();
+        for e in entries {
+            es.push(e?);
+        }
+        Ok(es)
+    }
+}
+
+
+// Private implementation
+impl Datastore {
     fn new_connection(path: &Path) -> Result<Connection> {
         let conn = Connection::open(path)?;
 
@@ -181,197 +395,12 @@ impl Datastore {
         }
         Ok(user_version)
     }
-
-    pub fn analyze(&self) -> Result<()> {
-        self.conn.execute("analyze", [])?;
-        Ok(())
-    }
-
-    pub fn record_matches<'a, T: IntoIterator<Item = &'a Match>>(
-        &mut self,
-        matches: T,
-    ) -> Result<usize> {
-        let _span = debug_span!("Datastore::record_matches", "{}", self.root_dir.display()).entered();
-
-        let tx = self.conn.transaction()?;
-        let mut stmt = tx.prepare_cached(indoc! {r#"
-            insert or replace into matches(
-                blob_id,
-                start_byte,
-                end_byte,
-                start_line,
-                start_column,
-                end_line,
-                end_column,
-                before_snippet,
-                matching_input,
-                after_snippet,
-                group_index,
-                group_input,
-                rule_name,
-                provenance_type,
-                provenance
-            ) values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        "#})?;
-        let mut num_changed = 0;
-        for m in matches {
-            let span = &m.location.offset_span;
-            let src = &m.location.source_span;
-            let (ptype, ppath) = match &m.provenance {
-                Provenance::File { path } => ("file", path.to_string_lossy()),
-                Provenance::GitRepo { path } => ("git", path.to_string_lossy()),
-            };
-            // FIXME: the number of changed rows is not the number of newly found matches!
-            num_changed += stmt.execute((
-                m.blob_id.hex(),
-                span.start,
-                span.end,
-                src.start.line,
-                src.start.column,
-                src.end.line,
-                src.end.column,
-                m.snippet.before.as_slice(),
-                m.snippet.matching.as_slice(),
-                m.snippet.after.as_slice(),
-                &m.capture_group_index,
-                m.match_content.as_slice(),
-                &m.rule_name,
-                ptype,
-                ppath,
-            ))?;
-        }
-        drop(stmt);
-        tx.commit()?;
-        Ok(num_changed)
-    }
-
-    pub fn summarize(&self) -> Result<MatchSummary> {
-        let _span = debug_span!("Datastore::summarize", "{}", self.root_dir.display()).entered();
-
-        let mut stmt = self.conn.prepare_cached(indoc! {r#"
-            select rule_name, count(*) grouped_count, sum(num_matches) total_count
-            from (
-                select group_input, rule_name, count(*) num_matches
-                from matches
-                group by 1, 2
-            )
-            group by 1
-            order by grouped_count desc
-        "#})?;
-        let entries = stmt.query_map((), |row| {
-            Ok(MatchSummaryEntry {
-                rule_name: row.get(0)?,
-                distinct_count: row.get(1)?,
-                total_count: row.get(2)?,
-            })
-        })?;
-        let mut es = Vec::new();
-        for e in entries {
-            es.push(e?);
-        }
-        Ok(MatchSummary(es))
-    }
-
-    pub fn root_dir(&self) -> &Path {
-        &self.root_dir
-    }
-
-    pub fn get_match_group_metadata(&self) -> Result<Vec<MatchGroupMetadata>> {
-        let _span =
-            debug_span!("Datastore::get_match_group_metadata", "{}", self.root_dir.display()).entered();
-
-        let mut stmt = self.conn.prepare_cached(indoc! {r#"
-            select group_input, rule_name, count(*)
-            from matches
-            group by 1, 2
-            order by 2
-        "#})?;
-        let entries = stmt.query_map((), |row| {
-            Ok(MatchGroupMetadata {
-                match_content: BString::new(row.get(0)?),
-                rule_name: row.get(1)?,
-                num_matches: row.get(2)?,
-            })
-        })?;
-        let mut es = Vec::new();
-        for e in entries {
-            es.push(e?);
-        }
-        Ok(es)
-    }
-
-    pub fn get_match_group_matches(
-        &self,
-        metadata: &MatchGroupMetadata,
-        limit: Option<usize>,
-    ) -> Result<Vec<Match>> {
-        let _span = debug_span!("Datastore::match_groups", "{}", self.root_dir.display()).entered();
-
-        let mut stmt = self.conn.prepare_cached(indoc! {r#"
-            select
-                blob_id,
-                start_byte,
-                end_byte,
-                start_line,
-                start_column,
-                end_line,
-                end_column,
-                before_snippet,
-                matching_input,
-                after_snippet,
-                group_index,
-                provenance_type,
-                provenance
-            from matches
-            where rule_name = ? and group_input = ?
-            order by blob_id, start_byte, end_byte
-            limit ?
-        "#})?;
-
-        let limit: i64 = match limit {
-            Some(limit) => limit.try_into().expect("limit should be convertible"),
-            None => -1,
-        };
-        let entries = stmt.query_map((&metadata.rule_name, metadata.match_content.as_slice(), limit), |row| {
-            let v0: String = row.get(0)?;
-            Ok(Match {
-                blob_id: BlobId::from_hex(&v0).expect("blob id from database should be valid"),
-                location: Location {
-                    offset_span: OffsetSpan {
-                        start: row.get(1)?,
-                        end: row.get(2)?,
-                    },
-                    source_span: SourceSpan {
-                        start: SourcePoint {
-                            line: row.get(3)?,
-                            column: row.get(4)?,
-                        },
-                        end: SourcePoint {
-                            line: row.get(5)?,
-                            column: row.get(6)?,
-                        },
-                    },
-                },
-                snippet: Snippet {
-                    before: BString::new(row.get(7)?),
-                    matching: BString::new(row.get(8)?),
-                    after: BString::new(row.get(9)?),
-                },
-                capture_group_index: row.get(10)?,
-                match_content: metadata.match_content.clone(),
-                rule_name: metadata.rule_name.clone(),
-                provenance: provenance_from_parts(row.get(11)?, row.get(12)?)
-                    .expect("provenance value from database should be valid"),
-            })
-        })?;
-        let mut es = Vec::new();
-        for e in entries {
-            es.push(e?);
-        }
-        Ok(es)
-    }
 }
 
+
+// -------------------------------------------------------------------------------------------------
+// Implementation Utilities
+// -------------------------------------------------------------------------------------------------
 fn provenance_from_parts(tag: String, path: String) -> Result<Provenance> {
     match tag.as_str() {
         "git" => Ok(Provenance::GitRepo {
@@ -384,10 +413,48 @@ fn provenance_from_parts(tag: String, path: String) -> Result<Provenance> {
     }
 }
 
+
+/// Get a path for a local clone of the given git URL underneath `root`.
+fn clone_destination(root: &std::path::Path, repo: &GitUrl) -> Result<std::path::PathBuf> {
+    Ok(root.join(repo.to_path_buf()))
+}
+
+#[cfg(test)]
+mod test {
+    macro_rules! clone_destination_success_tests {
+        ($($case_name:ident: ($root:expr, $repo:expr) => $expected:expr,)*) => {
+            mod clone_destination {
+                use crate::git_url::GitUrl;
+                use pretty_assertions::assert_eq;
+                use std::path::{PathBuf, Path};
+                use std::str::FromStr;
+                use super::super::clone_destination;
+
+                $(
+                    #[test]
+                    fn $case_name() {
+                        let expected: Option<PathBuf> = Some(Path::new($expected).to_owned());
+
+                        let root = Path::new($root);
+                        let repo = GitUrl::from_str($repo).expect("repo should be a URL");
+                        assert_eq!(clone_destination(root, &repo).ok(), expected);
+                    }
+                )*
+            }
+        }
+    }
+
+    clone_destination_success_tests! {
+        https_01: ("rel_root", "https://example.com/testrepo.git") => "rel_root/https/example.com/testrepo.git",
+        https_02: ("/abs_root", "https://example.com/testrepo.git") => "/abs_root/https/example.com/testrepo.git",
+    }
+}
+
 // -------------------------------------------------------------------------------------------------
 // MatchSummary
 // -------------------------------------------------------------------------------------------------
-/// A summary of matches in a `Datastore`
+
+/// A summary of matches in a `Datastore`.
 #[derive(Deserialize, Serialize)]
 pub struct MatchSummary(pub Vec<MatchSummaryEntry>);
 
@@ -410,6 +477,8 @@ impl std::fmt::Display for MatchSummary {
 // -------------------------------------------------------------------------------------------------
 // MatchGroupMetadata
 // -------------------------------------------------------------------------------------------------
+
+/// Metadata for a group of matches that have identical match content.
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MatchGroupMetadata {
     /// The name of the rule of all the matches in the group
